@@ -1,15 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import store from './store'
 import './assets/tailwind.css'
 
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  store,
-  render: function (h) { return h(App) }
-}).$mount('#app')
-
-window.Neutralino.init();
+window.Neutralino.init()
+createApp(App).use(store).use(router).mount('#app')
