@@ -1,9 +1,6 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { createStore } from 'vuex'
 
-Vue.use(Vuex)
-
-export default new Vuex.Store({
+export default createStore({
   state: {
     baseUrl: "",
     countries: [],
@@ -69,5 +66,5 @@ export default new Vuex.Store({
     volume(store, value) {
       store.volume = value
     }
-  },
+  }
 })
